feat(ui): TE-1270 moved notification channels view utils to separate file
