+++ thirdeye-ui/src/app/components/subscription-group-view/notification-channels-card/notification-channels-card.component.tsx
@@ -13,96 +13,12 @@
  * the License.
  */
 
-import { Icon } from "@iconify/react";
-import { Box, Grid } from "@material-ui/core";
-import React, { FunctionComponent, Key, useMemo } from "react";
+import { Grid } from "@material-ui/core";
+import React, { FunctionComponent, useMemo } from "react";
 import { useTranslation } from "react-i18next";
-import { lightV1 } from "../../../platform/utils";
-import {
-    NotificationSpec,
-    SpecType,
-} from "../../../rest/dto/subscription-group.interfaces";
 import { SubscriptionGroupViewCard } from "../subscription-group-view-card/subscription-group-view-card.component";
-import { SubscriptionGroupViewCardProps } from "../subscription-group-view-card/subscription-group-view-card.interface";
 import { NotificationChannelsCardProps } from "./notification-channels-card.interface";
-
-const getCardProps = (
-    channel: NotificationSpec,
-    t: (v: string) => string
-): SubscriptionGroupViewCardProps & { key?: Key } => {
-    const channelHeaderMap = {
-        [SpecType.Webhook]: t("label.webhook"),
-        [SpecType.Slack]: t("label.slack"),
-        [SpecType.EmailSendgrid]: t("label.email"),
-    };
-
-    const channelIconsMap = {
-        [SpecType.Webhook]: "material-symbols:webhook-rounded",
-        [SpecType.Slack]: "mdi:slack",
-        [SpecType.EmailSendgrid]: "ic:outline-email",
-    };
-
-    const cardProps: SubscriptionGroupViewCardProps & { key?: Key } = {
-        header: (
-            <Box alignItems="center" display="flex" gridGap={8}>
-                <Icon
-                    color={lightV1.palette.primary.dark}
-                    fontSize={32}
-                    icon={channelIconsMap[channel.type]}
-                />
-                {channelHeaderMap[channel.type] || t("label.email")}
-            </Box>
-        ),
-        rows: [],
-    };
-
-    switch (channel.type) {
-        case SpecType.Webhook:
-            cardProps.rows.push({
-                label: t("label.url"),
-                value: channel.params.url,
-            });
-
-            break;
-        case SpecType.Slack:
-            cardProps.rows.push({
-                label: t("label.url"),
-                value: channel.params.webhookUrl,
-            });
-
-            break;
-
-        case SpecType.EmailSendgrid:
-            cardProps.rows.push(
-                {
-                    label: t("label.sendgrid-api-key"),
-                    value: channel.params.apiKey,
-                },
-                {
-                    label: t("label.from"),
-                    value: channel.params.emailRecipients.from,
-                },
-                {
-                    label: t("label.to"),
-                    value: channel.params.emailRecipients.to.join(","),
-                }
-            );
-
-            break;
-        // This case below should NEVER happen, but adding here in case
-        // the back-end adds a new notification channel type without the
-        // UI getting updated to support the same
-        default:
-            cardProps.rows.push({
-                label: t("label.configuration"),
-                value: JSON.stringify(channel),
-            });
-
-            break;
-    }
-
-    return cardProps;
-};
+import { getCardProps } from "./notification-channels-card.utils";
 
 export const NotificationChannelsCard: FunctionComponent<NotificationChannelsCardProps> =
     ({ activeChannels }) => {
